+++ src/main/java/kr/co/chikenbreastsite/service/product/DeleteProductService.java
@@ -2,14 +2,9 @@ package kr.co.chikenbreastsite.service.product;
 
 import kr.co.chikenbreastsite.domain.dto.product.DeleteProductDto;
 import kr.co.chikenbreastsite.domain.entity.product.Product;
-import kr.co.chikenbreastsite.domain.entity.product.UserRole;
-import kr.co.chikenbreastsite.exception.BusinessLogicException;
 import kr.co.chikenbreastsite.exception.product.ProductDoesNotAlreadyExistException;
-import kr.co.chikenbreastsite.exception.product.ProductNotFoundException;
-import kr.co.chikenbreastsite.exception.users.DuplicationCellPhoneException;
 import kr.co.chikenbreastsite.exception.users.UsersNotRoleException;
-import kr.co.chikenbreastsite.exception.users.WrongPasswordException;
-import kr.co.chikenbreastsite.repository.product.ProductRepository;
+import kr.co.chikenbreastsite.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
