+++ src/main/java/kr/co/chikenbreastsite/service/product/GetProductService.java
@@ -3,12 +3,10 @@ package kr.co.chikenbreastsite.service.product;
 import kr.co.chikenbreastsite.domain.dto.product.GetProductDto;
 import kr.co.chikenbreastsite.domain.entity.product.Product;
 import kr.co.chikenbreastsite.exception.product.ProductNotFoundException;
-import kr.co.chikenbreastsite.repository.product.ProductRepository;
+import kr.co.chikenbreastsite.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
