+++ src/main/java/kr/co/chikenbreastsite/service/users/SignInService.java
@@ -5,11 +5,10 @@ import kr.co.chikenbreastsite.domain.entity.users.Users;
 import kr.co.chikenbreastsite.exception.ResponseFormat;
 import kr.co.chikenbreastsite.exception.users.UsersNotFoundException;
 import kr.co.chikenbreastsite.exception.users.WrongPasswordException;
-import kr.co.chikenbreastsite.repository.users.UsersRepository;
+import kr.co.chikenbreastsite.repository.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import javax.transaction.Transactional;
-import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
