+++ src/main/java/kr/co/chikenbreastsite/service/users/UsersDeleteService.java
@@ -4,7 +4,7 @@ import kr.co.chikenbreastsite.domain.dto.users.UserDeleteDto;
 import kr.co.chikenbreastsite.domain.entity.users.Users;
 import kr.co.chikenbreastsite.exception.users.UsersNotFoundException;
 import kr.co.chikenbreastsite.exception.users.WrongPasswordException;
-import kr.co.chikenbreastsite.repository.users.UsersRepository;
+import kr.co.chikenbreastsite.repository.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
