+++ src/main/java/kr/co/chikenbreastsite/service/users/UsersPasswordUpdateService.java
@@ -6,7 +6,7 @@ import kr.co.chikenbreastsite.exception.users.NewPasswordException;
 import kr.co.chikenbreastsite.exception.users.UsersNotFoundException;
 import kr.co.chikenbreastsite.exception.users.WrongNewCheckPasswordException;
 import kr.co.chikenbreastsite.exception.users.WrongPasswordException;
-import kr.co.chikenbreastsite.repository.users.UsersRepository;
+import kr.co.chikenbreastsite.repository.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
