+++ src/main/java/kr/co/chikenbreastsite/service/users/UsersUpdateService.java
@@ -3,7 +3,7 @@ package kr.co.chikenbreastsite.service.users;
 import kr.co.chikenbreastsite.domain.dto.users.UsersUpdateDto;
 import kr.co.chikenbreastsite.domain.entity.users.Users;
 import kr.co.chikenbreastsite.exception.users.*;
-import kr.co.chikenbreastsite.repository.users.UsersRepository;
+import kr.co.chikenbreastsite.repository.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
